+++ core/src/components/datetime/test/prefer-wheel/datetime.e2e.ts
@@ -40,6 +40,18 @@ test.describe('datetime: prefer wheel', () => {
         `datetime-wheel-time-date-diff-${page.getSnapshotSettings()}.png`
       );
     });
+    test('should render a condense header when specified', async ({ page }) => {
+      await page.setContent(`
+        <ion-datetime size="cover" presentation="time-date" prefer-wheel="true" value="2019-05-30T16:30:00"><div slot="title">My Custom Title</div></ion-datetime>
+      `);
+      await page.waitForSelector('.datetime-ready');
+
+      const datetime = page.locator('ion-datetime');
+
+      expect(await datetime.screenshot()).toMatchSnapshot(
+        `datetime-wheel-header-diff-${page.getSnapshotSettings()}.png`
+      );
+    });
   });
   test.describe('datetime: date wheel', () => {
     test.beforeEach(({ skip }) => {
@@ -136,6 +148,19 @@ test.describe('datetime: prefer wheel', () => {
           presentation="date"
           prefer-wheel="true"
         ></ion-datetime>
+
+        <script>
+          const mockToday = '2022-10-10T16:22';
+          Date = class extends Date {
+            constructor(...args) {
+              if (args.length === 0) {
+                super(mockToday)
+              } else {
+                super(...args);
+              }
+            }
+          }
+        </script>
       `);
 
       await page.waitForSelector('.datetime-ready');