+++ core/src/components/textarea/textarea-interface.ts
@@ -1,8 +1,14 @@
 export interface TextareaChangeEventDetail {
   value?: string | null;
+  event?: Event;
 }
 
-export interface TextareaCustomEvent extends CustomEvent {
-  detail: TextareaChangeEventDetail;
+export interface TextareaInputEventDetail {
+  value?: string | null;
+  event?: Event;
+}
+
+export interface TextareaCustomEvent<T = TextareaChangeEventDetail> extends CustomEvent {
+  detail: T;
   target: HTMLIonTextareaElement;
 }