+++ core/src/utils/framework-delegate.ts
@@ -53,6 +53,8 @@ export const CoreDelegate = () => {
     userComponentProps: any = {},
     cssClasses: string[] = []
   ) => {
+    const hasUserDefinedComponent = userComponent !== undefined;
+
     BaseComponent = parentElement;
     /**
      * If passing in a component via the `component` props
@@ -86,7 +88,7 @@ export const CoreDelegate = () => {
       BaseComponent.appendChild(el);
 
       await new Promise((resolve) => componentOnReady(el, resolve));
-    } else if (BaseComponent.children.length > 0) {
+    } else if (hasUserDefinedComponent && BaseComponent.children.length > 0) {
       // If there is no component, then we need to create a new parent
       // element to apply the css classes to.
       const el = BaseComponent.ownerDocument?.createElement('div');