+++ packages/react/src/components/IonLoading.tsx
@@ -1,9 +1,9 @@
-import { LoadingOptions, loadingController } from '@ionic/core/components';
+import { JSX } from '@ionic/core/components';
 import { defineCustomElement } from '@ionic/core/components/ion-loading.js';
 
-import { createControllerComponent } from './createControllerComponent';
+import { createInlineOverlayComponent } from './createInlineOverlayComponent';
 
-export const IonLoading = /*@__PURE__*/ createControllerComponent<
-  LoadingOptions,
+export const IonLoading = /*@__PURE__*/ createInlineOverlayComponent<
+  JSX.IonLoading,
   HTMLIonLoadingElement
->('ion-loading', loadingController, defineCustomElement);
+>('ion-loading', defineCustomElement);