+++ packages/react/test-app/tests/e2e/specs/overlay-components/IonLoading.cy.ts
@@ -9,7 +9,7 @@ describe('IonLoading', () => {
     cy.get('ion-loading').contains('Loading');
 
     //loading goes away after 1s
-    cy.get('ion-loading').should('not.exist');
+    cy.get('ion-loading').should('not.be.visible');
   });
 
   it('display loading and call dismiss to close it', () => {
@@ -17,7 +17,7 @@ describe('IonLoading', () => {
     cy.get('ion-button').contains('Show Loading, hide after 250 ms').click();
     cy.get('ion-loading').contains('Loading');
 
-    //verify loading is gone
-    cy.get('ion-loading').should('not.exist');
+    //verify loading is hidden
+    cy.get('ion-loading').should('not.be.visible');
   });
 });